// src/models/connect.d.ts
export type Effect = (
  action: AnyAction,
  effects: EffectsCommandMap & { select: <T>(func: (state: ConnectState) => T) => T },
) => void;

/**
 * @type P: Type of payload
 * @type C: Type of callback
 */
export type Dispatch = <P = any, C = (payload: P) => void>(action: {
  type: string;
  payload?: P;
  callback?: C;
  [key: string]: any;
}) => any;

export interface Loading {
  global: boolean;
  effects: { [key: string]: boolean | undefined };
  models: {
    global: boolean;
    menu: boolean;
    setting: boolean;
    user: boolean; // 当前登录用户
    users: boolean; // 用户管理
    role: boolean;
    permission: boolean;
    deviceProduct: boolean;
    deviceInstance: boolean;
    email: boolean;
    sms: boolean;
    mqttClient: boolean;
    ruleInstance: boolean;
    ruleModel: boolean;
    protocol: boolean;
    openApi: boolean;
    certificate: boolean;
    coapClient: boolean;
    httpClient: boolean;
    websocketClient: boolean;
    accessLogger: boolean;
    systemLogger: boolean;
    networkType: boolean;
    gateway: boolean;
    deviceGateway: boolean;
    noticeTemplate: boolean;
    noticeConfig: boolean;
    org: boolean;
    settings: boolean
  };
}

export interface ConnectState {
  login: LoginModelType;
  global: GlobalModelState;
  settings: SettingModelState;
  user: UserModelState;
  //以上是全局属性，以下是局部属性，局部属性应该是可选的但都不是 ?:
  loading?: Loading;
  users?: UsersModelState;
  role?: RoleModelState;
  permission?: PermissionModelState;
  deviceProduct?: DeviceProductModelState;
  deviceInstance?: DeviceInstanceModelState;
  email?: EmailModelState;
  sms?: SmsModelState;
  mqttClient?: MqttClientModelState;
  ruleInstance?: RuleInstanceModelState;
  ruleModel?: RuleModelModelState;
  protocol?: ProtocolModelState;
  openApi?: OpenApiModelState;
  certificate?: CertificateModelState;
  coapClient?: CoapClientModelState;
  httpClient?: HttpClientModelState;
  websocketClient?: WebsocketClientModelState;
  accessLogger?: AccessLoggerModelState;
  systemLogger?: SystemLoggerModelState;
  networkType?: NetworkTypeState;
  gateway?: GatewayModelState;
  deviceGateway?: DeviceGatewayState;
  noticeTemplate?: NoticeTemplateState;
  noticeConfig?: NoticeConfigState;
  org?: OrgModelState;
  //settings: SettingModelState;//两个settings？？
  xxxxx:boolean //ts声明没有生效？？环境配置哪里有问题？？？
}